feat(leadspace-interior): move component to leadspaces directory
