+++ src/index.ts
@@ -1,3 +1,4 @@
 export { Eyebrow } from './components/eyebrow';
 export { AlternatingFeatureWithImage } from './components/alternatingFeatureWithImage';
 export { StatsCarousel, StatsCarouselSlide } from './components/statsCarousel';
+export { LeadspaceInterior } from './components/leadspaces/leadspaceInterior';
